+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [6.13.1] 2024-10-17
+
+### Fixed
+
+- [TD-6894] Changeset validations overwrite existing content 
+
 ## [6.13.0] 2024-10-15
 
 ### Added